requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
alembic==1.13.2

# Pydantic & Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# LLM Providers
google-generativeai==0.8.2

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
uuid7==0.1.0

# Security
passlib==1.7.4
python-jose[cryptography]==3.3.0
